# Minimize unused memory allocated in DataChannel.readLoop.