--- a/datachannel_test.go
+++ b/datachannel_test.go
@@ -1,6 +1,7 @@
 package webrtc
 
 import (
+	"fmt"
 	"io"
 	"sync"
 	"testing"
@@ -15,7 +16,7 @@ import (
 // bindings this is a requirement).
 const expectedLabel = "data"
 
-func closePairNow(t *testing.T, pc1, pc2 io.Closer) {
+func closePairNow(t testing.TB, pc1, pc2 io.Closer) {
 	var fail bool
 	if err := pc1.Close(); err != nil {
 		t.Errorf("Failed to close PeerConnection: %v", err)
@@ -61,6 +62,52 @@ func closeReliabilityParamTest(t *testing.T, pc1, pc2 *PeerConnection, done chan
 	}
 
 	closePair(t, pc1, pc2, done)
+}
+
+func BenchmarkDataChannelSend2(b *testing.B)  { benchmarkDataChannelSend(b, 2) }
+func BenchmarkDataChannelSend4(b *testing.B)  { benchmarkDataChannelSend(b, 4) }
+func BenchmarkDataChannelSend8(b *testing.B)  { benchmarkDataChannelSend(b, 8) }
+func BenchmarkDataChannelSend16(b *testing.B) { benchmarkDataChannelSend(b, 16) }
+func BenchmarkDataChannelSend32(b *testing.B) { benchmarkDataChannelSend(b, 32) }
+
+// See https://github.com/pion/webrtc/issues/1516
+func benchmarkDataChannelSend(b *testing.B, numChannels int) {
+	offerPC, answerPC, err := newPair()
+	if err != nil {
+		b.Fatalf("Failed to create a PC pair for testing")
+	}
+
+	open := make(map[string]chan bool)
+	answerPC.OnDataChannel(func(d *DataChannel) {
+		if _, ok := open[d.Label()]; !ok {
+			// Ignore anything unknown channel label.
+			return
+		}
+		d.OnOpen(func() { open[d.Label()] <- true })
+	})
+
+	var wg sync.WaitGroup
+	for i := 0; i < numChannels; i++ {
+		label := fmt.Sprintf("dc-%d", i)
+		open[label] = make(chan bool)
+		wg.Add(1)
+		dc, err := offerPC.CreateDataChannel(label, nil)
+		assert.NoError(b, err)
+
+		dc.OnOpen(func() {
+			<-open[label]
+			for n := 0; n < b.N/numChannels; n++ {
+				if err := dc.SendText("Ping"); err != nil {
+					b.Fatalf("Unexpected error sending data (label=%q): %v", label, err)
+				}
+			}
+			wg.Done()
+		})
+	}
+
+	assert.NoError(b, signalPair(offerPC, answerPC))
+	wg.Wait()
+	closePairNow(b, offerPC, answerPC)
 }
 
 func TestDataChannel_Open(t *testing.T) {