@bshimc 2020-11-12
@Sean-Der I've addressed all the CI issues I could but I'm not sure how the system is calculating the reduced coverage.  Perhaps it's that Benchmark tests aren't run.