@Sean-Der 2020-11-12
@bshimc We unfortunately don't have any benchmark infrastructure. It would be really awesome if we did though! Do you have anything to automate catching memory/performance regressions?  

I am fine with with merging this as is, and we can come back to it. If you are using Pion I would love to add you to the organization and get your help creating a better 'perf culture'. That and fuzz testing are our biggest misses right now.