@jsmouret 2020-11-12
The original patch in #1516 was a lot nicer without the extra copy.
As v3 is on the way and already has plenty of API changes like the Track interface,
could we add the requirement that OnMessage buffer be only used inside the callback?