@Sean-Der 2020-11-12
Lets do it! @jsmouret  would you mind adding an entry to the 'v3.0.0' release notes for it?

Would you mind doing the original patch @bshimc? Sorry for the churn, I tend to be more cautious but I love seeing community excitement :) 

I still would love to see if we could get some CI and start paying down performance debt.