@bshimc 2020-11-12
@Sean-Der @jsmouret I appreciate the enthusiasm but I'd recommend holding off on making the aggressive change.  On further reflection the original patch supplied in the issue is poor API design that is begging to be accidentally misused.  The fundamental problem is that we're passing byte buffers around.  Performance sensitive APIs generally pass io.ReadCloser objects around to avoid allocations so I tried taking a page out of the http.Request.Body API where it's the user's responsibility to call .Close().  Most importantly, this allows the user to decide what the lifecycle of the allocated memory should be.  See https://github.com/pion/webrtc/compare/master...bshimc:readcloser for a working example.  Benchmarking that yielded results that suggest that my benchmark is flawed; there are basically no discernable differences between what's shipping now and the rlBufPool implementation in that branch.  I'll try to run a distributed load test to see if there are improvements because I no longer trust my synthetic Benchmark test. This is all along winded way of advising caution.  There are some gaps in my understanding of what's going on that I need to fill.

```diff
diff --git a/datachannelmessage.go b/datachannelmessage.go
index 1e3c63b..7cbafdf 100644
--- a/datachannelmessage.go
+++ b/datachannelmessage.go
@@ -1,10 +1,15 @@
 package webrtc

+import "io"
+
 // DataChannelMessage represents a message received from the
 // data channel. IsString will be set to true if the incoming
 // message is of the string type. Otherwise the message is of
 // a binary type.
 type DataChannelMessage struct {
        IsString bool
-       Data     []byte
+
+       // It is the caller's responsibility to close Data to release memory
+       // resources.
+       Data io.ReadCloser
 }
diff --git a/datachannel.go b/datachannel.go
index ba1269d..d44509d 100644
--- a/datachannel.go
+++ b/datachannel.go
@@ -3,6 +3,7 @@
 package webrtc

 import (
+       "bytes"
        "errors"
        "fmt"
        "io"
@@ -44,7 +45,7 @@ type DataChannel struct {
        // "blob". This attribute controls how binary data is exposed to scripts.
        // binaryType                 string

-       onMessageHandler    func(DataChannelMessage)
+       onMessageHandler    func(*DataChannelMessage)
        openHandlerOnce     sync.Once
        onOpenHandler       func()
        onCloseHandler      func()
@@ -260,13 +261,13 @@ func (d *DataChannel) onClose() {
 // in size. Check out the detach API if you want to use larger
 // message sizes. Note that browser support for larger messages
 // is also limited.
-func (d *DataChannel) OnMessage(f func(msg DataChannelMessage)) {
+func (d *DataChannel) OnMessage(f func(msg *DataChannelMessage)) {
        d.mu.Lock()
        defer d.mu.Unlock()
        d.onMessageHandler = f
 }

-func (d *DataChannel) onMessage(msg DataChannelMessage) {
+func (d *DataChannel) onMessage(msg *DataChannelMessage) {
        d.mu.RLock()
        handler := d.onMessageHandler
        d.mu.RUnlock()
@@ -311,11 +312,33 @@ func (d *DataChannel) onError(err error) {
        }
 }

+// See https://github.com/pion/webrtc/issues/1516
+// nolint:gochecknoglobals
+var rlBufPool sync.Pool
+
+type rlBuf struct {
+       *bytes.Reader
+       b [dataChannelBufferSize]byte
+}
+
+func (r *rlBuf) Close() error {
+       r.Reader = nil
+       rlBufPool.Put(r)
+       return nil
+}
+
 func (d *DataChannel) readLoop() {
        for {
-               buffer := make([]byte, dataChannelBufferSize)
-               n, isString, err := d.dataChannel.ReadDataChannel(buffer)
+               var b *rlBuf
+               if v := rlBufPool.Get(); v != nil {
+                       b = v.(*rlBuf)
+               } else {
+                       b = &rlBuf{}
+               }
+
+               n, isString, err := d.dataChannel.ReadDataChannel(b.b[:])
                if err != nil {
+                       rlBufPool.Put(b)
                        d.setReadyState(DataChannelStateClosed)
                        if err != io.EOF {
                                d.onError(err)
@@ -324,7 +347,8 @@ func (d *DataChannel) readLoop() {
                        return
                }

-               d.onMessage(DataChannelMessage{Data: buffer[:n], IsString: isString})
+               b.Reader = bytes.NewReader(b.b[:n])
+               d.onMessage(&DataChannelMessage{Data: b, IsString: isString})
        }
 }

```


> If you are using Pion I would love to add you to the organization

Sure; I'm not able to make any concrete time or resource commitments (and I'm certainly not a domain expert in this area) but I'm happy to help where possible.  Feel free to add my alter ego (@bshi) to the organization.