@Sean-Der 2020-11-14
@bshimc 

The perf improvement here is really exciting! Check out `examples/data-channels-flow-controls` 

**before**
```
2020/11/14 14:46:31 Throughput: 144.323 Mbps
2020/11/14 14:46:32 Throughput: 147.815 Mbps
2020/11/14 14:46:33 Throughput: 147.522 Mbps
2020/11/14 14:46:34 Throughput: 147.551 Mbps
2020/11/14 14:46:35 Throughput: 147.756 Mbps
2020/11/14 14:46:36 Throughput: 145.140 Mbps
2020/11/14 14:46:37 Throughput: 144.695 Mbps
2020/11/14 14:46:38 Throughput: 144.200 Mbps
2020/11/14 14:46:39 Throughput: 144.131 Mbps
2020/11/14 14:46:40 Throughput: 144.841 Mbps
2020/11/14 14:46:41 Throughput: 145.251 Mbps
2020/11/14 14:46:42 Throughput: 145.237 Mbps
```

**after**
```
2020/11/14 14:44:53 Throughput: 204.586 Mbps
2020/11/14 14:44:54 Throughput: 206.091 Mbps
2020/11/14 14:44:55 Throughput: 205.737 Mbps
2020/11/14 14:44:56 Throughput: 205.667 Mbps
2020/11/14 14:44:57 Throughput: 206.387 Mbps
2020/11/14 14:44:58 Throughput: 207.378 Mbps
2020/11/14 14:44:59 Throughput: 208.151 Mbps
2020/11/14 14:45:00 Throughput: 208.919 Mbps
2020/11/14 14:45:01 Throughput: 209.053 Mbps
2020/11/14 14:45:02 Throughput: 209.243 Mbps
2020/11/14 14:45:03 Throughput: 209.546 Mbps
2020/11/14 14:45:04 Throughput: 210.056 Mbps
2020/11/14 14:45:05 Throughput: 210.680 Mbps
2020/11/14 14:45:06 Throughput: 210.781 Mbps
2020/11/14 14:45:07 Throughput: 211.216 Mbps
```